requirements.txt
requests
beautifulsoup4
pyodbc
python-dotenv
tenacity
